+++ app/src/main/AndroidManifest.xml
@@ -24,6 +24,10 @@
             android:name=".userreg"
             android:theme="@style/AppTheme">
         </activity>
+        <activity
+            android:name=".reservation"
+            android:theme="@style/AppTheme">
+        </activity>
     </application>
 
 </manifest>
