+++ app/src/main/java/first/com/MainActivity.java
@@ -22,51 +22,26 @@ import cz.msebera.android.httpclient.Header;
 
 
 public class MainActivity extends AppCompatActivity {
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.userlist);
 
-    private ArrayList  data = new ArrayList<String>();
-    private void loadData(){
-        String urlString = "https://api.airtable.com/v0/appgPqAWrw2xTWKdx/List of divisions?api_key=keyKStB2L0gi9sY2B";
-        AsyncHttpClient client = new AsyncHttpClient();
-        client.get(urlString, new JsonHttpResponseHandler() {
-            public void onSuccess(int statusCode,Header[] headers, JSONObject response) {
-                Toast.makeText(getApplicationContext(),
-                        "Success!", Toast.LENGTH_LONG).show();
-                Log.d("Hot Text:", response.toString());
-                ListView kindview =(ListView)findViewById(R.id.kindview);
-                TextView tview =(TextView)findViewById(R.id.tview);
-                String Json = response.toString();
-                try {
-                    JSONArray Array = response.getJSONArray("records");
-                    for(int i = 0; i<Array.length();i++){
-                        JSONObject userdata = Array.getJSONObject(i);
-                        JSONObject fields=userdata.getJSONObject("fields");
-                        String id = fields.getString("Divisions_name");
-                        data.add(id);
-                        ArrayAdapter<String> adapter=new ArrayAdapter<String>(MainActivity.this,android.R.layout.simple_list_item_1,data);
-                        ListView listview=(ListView)findViewById(R.id.kindview);
-                        listview.setAdapter(adapter);
-                    }
-                } catch (JSONException e) {
-                    e.printStackTrace();
-                }
-
-            }
-
-
-            public void onFailure(int statusCode, Header[] headers, Throwable e, JSONObject error) {
-                Toast.makeText(getApplicationContext(),
-                        "Error: " + statusCode + " " + e.getMessage(),
-                        Toast.LENGTH_LONG).show();
-                // Log error message
-                Log.e("Hot Text:", statusCode + " " + e.getMessage());
+        Button nextPageBtn = (Button)findViewById(R.id.reservation);
+        nextPageBtn.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                Intent intent = new Intent();
+                intent.setClass(MainActivity.this , reservation.class);
+                startActivity(intent);
             }
         });
     }
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.reservation);
-        loadData();
-    }
+    /*
+    //reservation
+
+
+    */
+    //activity_main
     /*private void loadData(){
         String urlString = "https://api.airtable.com/v0/appgPqAWrw2xTWKdx/User?api_key=keyUwcLvTO51TNEHV";
         AsyncHttpClient client = new AsyncHttpClient();
