+++ app/src/main/res/layout/reservation.xml
@@ -72,8 +72,10 @@
 
     <ListView
         android:id="@+id/kindview"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent" />
+        android:layout_width="383dp"
+        android:layout_height="match_parent"
+        android:layout_marginHorizontal="15dp"
+        android:background="#ffffff"/>
     <TextView
         android:id="@+id/tview"
         android:layout_width="match_parent"
