+++ app/src/main/res/layout/userlist.xml
@@ -50,6 +50,7 @@
             android:layout_height="180dp"
             android:orientation="horizontal">
             <Button
+                android:id="@+id/reservation"
                 android:layout_width="0dp"
                 android:layout_marginLeft="30dp"
                 android:layout_marginRight="12dp"
